SQL Table changes
Renamed book to reference_book since it's a subset of reference in the
end. Changed the modified column of reference to auto update, since
that's the point of modified. Removed auto_increment, it's not supposed
to be in these scripts at all, should default to zero. Renamed privacy
row in content, to content_privacy_id...which it should have been
before.